lab-assistant-backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
google-generativeai
PyPDF2
python-docx
python-dotenv
